Hi Koddy,
The lambda expression you wrote for the comparator was fine except for attempting to access the private string field of the token arguments directly; while a method called on an instance can access *that instance's private fields directly (through `this`, or just by naming the field [sometimes naming `this` is unnecessary or "bad form" according to some engineers]), you cannot do the same when the method is a static method.

Static methods can access private (and of course public) static fields, because they're owned by the class, but they cannot access instance fields which are private.